+++ Resources/Luka_Backlace/Editor/Project.cs
@@ -98,7 +98,10 @@ namespace Luka.Backlace
         {
             new CustomBadge("_AnimeMode", new Dictionary<float, string> {
                 { 0f, "custom_badge_ramp" },
-                { 1f, "custom_badge_procedural" },
+                { 1f, "custom_badge_halftone" },
+                { 2f, "custom_badge_hifi" },
+                { 3f, "custom_badge_skin" },
+                { 4f, "custom_badge_wrapped" },
             }, new Color(0.8f, 0.2f, 0.3f))
         };
         public static readonly List<CustomBadge> blend_mode_badges = new List<CustomBadge>